Remove the workaround for short filter names
This workaround was added when the nextVisit of LATISS observations
came as short filter names, such as "r".

Since mid-Oct 2023, the nextVisit of LATISS observations come
with the full "{filter}~{grating}" name, such as "SDSSr_65mm~empty",
and this temporary workaround is no longer needed. Other instruments
have different filters and this pattern does not apply either.