Accept empty filters with suffix in their names
Empty filter names from the summit can come with a suffix, and
such suffix is ignored in the stack and replaced by simply
"empty" in obs_lsst.

As we use this filter name to directly query for calibration and
other datasets in the butler, the name is modified here so to
match the filter naming in the butler.

This workaround can be removed when all summit scripts we care
follow the stack's filter naming.