Ensure "in process" metric is always decremented.
The previous code would leave a request marked as "in process" even if
the service, for example, timed out waiting for a response. This
introduced a permanent offset into the metric that could only be
cleared by rebooting the fan-out service. The new code is
exception-safe and should not have this problem.