Remove special-casing of LSSTComCam/LSSTCam.
The original behavior was to crash and restart the entire fan-out
service if it got a message it didn't know how to handle. If the
message loop instead catches the corresponding exception, then there is
no need to specifically exempt the LSST cameras.

With this commit, it's now safe to configure the SUPPORTED_INSTRUMENTS
env variable.